refactor(revset): move `resolve_commits` into `revset` module
